# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vue Laravel Template API",
        "description": "API documentation for the Vue Laravel Template project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get all blog posts",
                "operationId": "9819e5802278e060898795fe3c919599",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language code (e.g., en, ms)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of blog posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BlogListItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create a new blog post",
                "operationId": "514d226035b3d74b6ccf9837b3d1ed36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "language_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "My New Blog",
                                        "maxLength": 200
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Full content of the blog post..."
                                    },
                                    "language_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogStoreResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blogs/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get a single blog post by ID",
                "operationId": "edf2058da1858ee867ebc4789b3195f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language code (e.g., en, ms)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog post details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update a blog post",
                "operationId": "6a99affa8f7d94cd429d58eb73cf0e88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title",
                                        "maxLength": 200
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content..."
                                    },
                                    "language_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - insufficient permissions"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/blogs/{id}/comment": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Add a comment to a blog post",
                "operationId": "99c3741e42ab022a75f252d254ae804e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Great article!"
                                    },
                                    "reply_to_id": {
                                        "description": "ID of the parent comment for replies",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BlogListItem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "My Blog Post"
                    },
                    "summary": {
                        "type": "string",
                        "example": "This is a summary..."
                    },
                    "comment_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "reaction_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "author": {
                        "type": "string",
                        "format": "email",
                        "example": "author@example.com"
                    },
                    "published_at": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "BlogDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "My Blog Post"
                    },
                    "content": {
                        "type": "string",
                        "example": "Full content of the blog post..."
                    },
                    "author": {
                        "type": "string",
                        "format": "email",
                        "example": "author@example.com"
                    },
                    "published_at": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z",
                        "nullable": true
                    },
                    "reaction_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "sections": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BlogSection"
                        }
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BlogComment"
                        }
                    }
                },
                "type": "object"
            },
            "BlogSection": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order": {
                        "type": "integer",
                        "example": 0
                    },
                    "image": {
                        "type": "string",
                        "example": "http://localhost:8000/storage/images/photo.jpg",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "example": "Section Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Section content..."
                    }
                },
                "type": "object"
            },
            "BlogComment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "type": "string",
                        "format": "email",
                        "example": "commenter@example.com"
                    },
                    "content": {
                        "type": "string",
                        "example": "Great post!"
                    },
                    "reply_to_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "MessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Successfully logged out"
                    }
                },
                "type": "object"
            },
            "CommentResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Comment appended successfully."
                    },
                    "comment": {
                        "$ref": "#/components/schemas/BlogComment"
                    }
                },
                "type": "object"
            },
...
            "BlogStoreResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Blog structural draft created successfully!"
                    },
                    "blog_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}